added port mappings & cache
